backend/requirements.txt
# FAM Backend Requirements
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0

# Database
# SQLite is built-in to Python

# Web Scraping
beautifulsoup4>=4.12.0
aiohttp>=3.8.0
lxml>=4.9.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
